Adding Entry point for New onboarding with valid values to begin with
