+++ ios/brave-ios/Sources/Brave/Frontend/Browser/BrowserViewController/BVC+Onboarding.swift
@@ -20,15 +20,22 @@ extension BrowserViewController {
   func presentOnboardingIntro() {
     if Preferences.DebugFlag.skipOnboardingIntro == true { return }
 
-    Preferences.AppState.isOnboardingActive.value = true
-    presentOnboardingWelcomeScreen(on: self)
+    // If locale is JP, start the new onboarding process
+    // This will be the case as long as new onboarding is active for JAPAN
+    if Locale.current.regionCode == "JP" {
+      presentFocusOnboarding()
+    } else {
+      presentOnboardingWelcomeScreen()
+    }
   }
 
-  private func presentOnboardingWelcomeScreen(on parentController: UIViewController) {
+  // MARK: Welcome Onboarding - Regions except JAPAN
+
+  private func presentOnboardingWelcomeScreen() {
     // 1. Existing user.
     // 2. User already completed onboarding.
     if Preferences.Onboarding.basicOnboardingCompleted.value == OnboardingState.completed.rawValue {
-      Preferences.AppState.isOnboardingActive.value = false
+      Preferences.AppState.shouldDeferPromotedPurchase.value = false
       return
     }
 
@@ -42,13 +49,13 @@ extension BrowserViewController {
         attributionManager: attributionManager
       )
       onboardingController.modalPresentationStyle = .fullScreen
-      parentController.present(onboardingController, animated: false)
+      present(onboardingController, animated: false)
       isOnboardingOrFullScreenCalloutPresented = true
     }
   }
 
   func showNTPOnboarding() {
-    Preferences.AppState.isOnboardingActive.value = false
+    Preferences.AppState.shouldDeferPromotedPurchase.value = false
     iapObserver.savedPayment = nil
 
     if !topToolbar.inOverlayMode,
@@ -95,7 +102,7 @@ extension BrowserViewController {
         guard let self = self else { return }
 
         Preferences.FullScreenCallout.omniboxCalloutCompleted.value = true
-        Preferences.AppState.isOnboardingActive.value = false
+        Preferences.AppState.shouldDeferPromotedPurchase.value = false
 
         self.triggerPromotedInAppPurchase(savedPayment: self.iapObserver.savedPayment)
       },
@@ -103,7 +110,7 @@ extension BrowserViewController {
         guard let self = self else { return }
 
         Preferences.FullScreenCallout.omniboxCalloutCompleted.value = true
-        Preferences.AppState.isOnboardingActive.value = false
+        Preferences.AppState.shouldDeferPromotedPurchase.value = false
 
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.25) {
           self.topToolbar.tabLocationViewDidTapLocation(self.topToolbar.locationView)
@@ -116,7 +123,10 @@ extension BrowserViewController {
     let basicOnboardingNotCompleted =
       Preferences.Onboarding.basicOnboardingProgress.value != OnboardingProgress.newTabPage.rawValue
 
-    if basicOnboardingNotCompleted, showNTPEducation().isEnabled, let url = showNTPEducation().url {
+    // NTP Education screen should load after onboarding is finished and user is on locale JP
+    let (educationPermitted, url) = (Locale.current.regionCode == "JP", URL.brave.ntpTutorialPage)
+
+    if basicOnboardingNotCompleted, educationPermitted {
       tabManager.addTab(
         PrivilegedRequest(url: url) as URLRequest,
         afterTab: self.tabManager.selectedTab,
@@ -194,15 +204,9 @@ extension BrowserViewController {
     )
   }
 
-  /// New Tab Page Education screen should load after onboarding is finished and user is on locale JP
-  /// - Returns: A tuple which shows NTP Education is enabled and URL to be loaded
-  func showNTPEducation() -> (isEnabled: Bool, url: URL?) {
-    return (Locale.current.regionCode == "JP", .brave.ntpTutorialPage)
-  }
-
   func completeOnboarding(_ controller: UIViewController) {
     Preferences.Onboarding.basicOnboardingCompleted.value = OnboardingState.completed.rawValue
-    Preferences.AppState.isOnboardingActive.value = false
+    Preferences.AppState.shouldDeferPromotedPurchase.value = false
     controller.dismiss(animated: true)
   }
 
@@ -221,7 +225,11 @@ extension BrowserViewController {
     popover.arrowDistance = 10.0
 
     // Create a border / placeholder view
-    let borderView = BorderView(frame: frame, cornerRadius: cornerRadius, colouredBorder: true)
+    let borderView = NotificationBorderView(
+      frame: frame,
+      cornerRadius: cornerRadius,
+      colouredBorder: true
+    )
     let placeholderView = UIView(frame: frame).then {
       $0.alpha = 0.0
       $0.frame = frame
@@ -294,35 +302,18 @@ extension BrowserViewController {
   }
 }
 
-// MARK: BorderView
+extension BrowserViewController {
 
-private class BorderView: UIView {
+  // MARK: Day 0 Focus Onboarding - JAPAN Region
 
-  public var didClickBorderedArea: (() -> Void)?
+  func presentFocusOnboarding() {
 
-  init(frame: CGRect, cornerRadius: CGFloat, colouredBorder: Bool = false) {
-    let borderLayer = CAShapeLayer().then {
-      let frame = frame.with { $0.origin = .zero }
-      $0.strokeColor = colouredBorder ? UIColor.braveLighterBlurple.cgColor : UIColor.white.cgColor
-      $0.fillColor = UIColor.clear.cgColor
-      $0.lineWidth = 2.0
-      $0.strokeEnd = 1.0
-      $0.path = UIBezierPath(roundedRect: frame, cornerRadius: cornerRadius).cgPath
+    // Check user has never seen onboarding - new user
+    guard Preferences.Onboarding.basicOnboardingCompleted.value == OnboardingState.unseen.rawValue
+    else {
+      Preferences.AppState.shouldDeferPromotedPurchase.value = false
+      return
     }
 
-    super.init(frame: frame)
-    layer.addSublayer(borderLayer)
-
-    addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(onClickBorder(_:))))
-  }
-
-  @available(*, unavailable)
-  required init(coder: NSCoder) {
-    fatalError()
-  }
-
-  @objc
-  private func onClickBorder(_ tap: UITapGestureRecognizer) {
-    didClickBorderedArea?()
   }
 }
